Move source detection to new module
